+++ sync-for-expenses/Codat/Sync/Expenses/Models/Components/PhoneNumberItems.cs
@@ -9,12 +9,23 @@
 #nullable enable
 namespace Codat.Sync.Expenses.Models.Components
 {
+    using Codat.Sync.Expenses.Models.Components;
     using Codat.Sync.Expenses.Utils;
+    using Newtonsoft.Json;
     
-    /// <summary>
-    /// A collection of user-defined key-value pairs that store custom metadata against the company.
-    /// </summary>
-    public class Tags
+    public class PhoneNumberItems
     {
+
+        /// <summary>
+        /// A phone number.
+        /// </summary>
+        [JsonProperty("number")]
+        public string? Number { get; set; } = null;
+
+        /// <summary>
+        /// The type of phone number
+        /// </summary>
+        [JsonProperty("type")]
+        public PhoneNumberType Type { get; set; } = default!;
     }
 }