+++ sync-for-expenses/Codat/Sync/Expenses/Models/Components/ReimbursableExpenseTransaction.cs
@@ -32,6 +32,9 @@ namespace Codat.Sync.Expenses.Models.Components
         [JsonProperty("contactRef")]
         public ReimbursementContactRef ContactRef { get; set; } = default!;
 
+        [JsonProperty("apAccountRef")]
+        public ApAccountRef? ApAccountRef { get; set; } = null;
+
         /// <summary>
         /// In Codat&apos;s data model, dates and times are represented using the &lt;a class=&quot;external&quot; href=&quot;https://en.wikipedia.org/wiki/ISO_8601&quot; target=&quot;_blank&quot;&gt;ISO 8601 standard&lt;/a&gt;. Date and time fields are formatted as strings; for example:<br/>
         ///